+++ learningTools/fio.h
@@ -21,6 +21,6 @@ typedef struct{
 	int right;
 }TreeNode;
 
-UCHAR *DecodeBookData(INT32 *decodeDataLen_p, char *filename);
-UCHAR *DecodeEvalData(INT32 *decodeDataLen_p, char *filename);
+UCHAR *DecodeBookData(size_t *decodeDataLen_p, char *filename);
+UCHAR *DecodeEvalData(size_t *decodeDataLen_p, char *filename);
 BOOL OpenMpcInfoData(char *filename);
